+++ changelog.md
@@ -1,2 +1,2 @@
 ### Tesseract 1.0.35
-- Initial release of Tesseract for NeoForge
+- Initial release of Tesseract for Minecraft 1.20.5 & 1.20.6